Fix init.sh before running init script

+++ jekyll-theme-chirpy-7.2.4/jekyll-theme-chirpy-7.2.4/tools/init.sh
@@ -66,7 +66,10 @@ check_env() {
 
 reset_latest() {
   git reset --hard "$RELEASE_HASH"
+
+if [[ -n $(git clean -nd) ]]; then
   git clean -fd
+fi
   git submodule update --init --recursive
 }
